CRUD operations implementation

+++ ads-dental-surgeries-lab7/src/main/java/edu/miu/cs/cs489appsd/adsdentalsurgerieslab7/services/impl/PatientServiceImpl.java
@@ -9,6 +9,7 @@ import edu.miu.cs.cs489appsd.adsdentalsurgerieslab7.services.PatientService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -22,6 +23,7 @@ public class PatientServiceImpl implements PatientService {
     @Override
     public PatientDto addPatient(PatientDto patientDto) {
         Address address = Address.builder()
+                .id(patientDto.addressDto().id())
                 .street(patientDto.addressDto().street())
                 .city(patientDto.addressDto().city())
                 .state(patientDto.addressDto().state())
@@ -40,13 +42,41 @@ public class PatientServiceImpl implements PatientService {
     }
 
     @Override
-    public Patient updatePatient(Patient patient) {
-        return patientRepository.saveAndFlush(patient);
+    public PatientDto updatePatient(String patientId, PatientDto patientDto) {
+        Patient patient = patientRepository.findById(patientId).orElse(null);
+        if (Objects.nonNull(patient)) {
+            Address address = Address.builder()
+                    .id(patientDto.addressDto().id())
+                    .street(patientDto.addressDto().street())
+                    .city(patientDto.addressDto().city())
+                    .state(patientDto.addressDto().state())
+                    .zip(patientDto.addressDto().zip())
+                    .build();
+            patient.setId(patientId);
+            patient.setFirstName(patientDto.firstName());
+            patient.setLastName(patientDto.lastName());
+            patient.setEmail(patientDto.email());
+            patient.setPhoneNumber(patientDto.phoneNumber());
+            patient.setDateOfBirth(patientDto.dateOfBirth());
+            patient.setAddress(address);
+        }
+        assert patient != null;
+        Patient updatedPatient = patientRepository.saveAndFlush(patient);
+
+        AddressDto addressDto = new AddressDto(updatedPatient.getAddress().getId(),
+                updatedPatient.getAddress().getStreet(), updatedPatient.getAddress().getCity(),
+                updatedPatient.getAddress().getState(), updatedPatient.getAddress().getZip());
+
+        return new PatientDto(updatedPatient.getId(), updatedPatient.getFirstName(), updatedPatient.getLastName(),
+                updatedPatient.getEmail(), updatedPatient.getPhoneNumber(), updatedPatient.getDateOfBirth(), addressDto);
     }
 
     @Override
-    public void deletePatient(Patient patient) {
-        patientRepository.delete(patient);
+    public void deletePatient(String patientId) {
+        Patient patient = patientRepository.findById(patientId).orElse(null);
+        if (Objects.nonNull(patient)) {
+            patientRepository.delete(patient);
+        }
     }
 
     @Override
@@ -70,6 +100,5 @@ public class PatientServiceImpl implements PatientService {
                     patient.getPhoneNumber(), patient.getDateOfBirth(), addressDto);
         }).orElse(new PatientDto(null, null, null, null, null, null, null));
     }
-
 
 }
